requirements.txt
aggdraw==1.3.19
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
click==8.2.1
docopt==0.6.2
fastapi==0.116.1
h11==0.16.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
psd-tools==1.10.8
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
scikit-image==0.25.2
scipy==1.16.0
sniffio==1.3.1
starlette==0.47.2
tifffile==2025.6.11
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
